Minor Camera refactor
