+++ src/components/index.ts
@@ -1,3 +1,3 @@
-export { default as CameraFeed, CameraStatus } from './CameraFeed';
+export { default as Camera, CameraStatus } from './Camera';
 export { default as SudokuScanner } from './SudokuScanner';
 export { default as SudokuSolver } from './SudokuSolver';
